requirements.txt
APScheduler==3.10.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msmart==0.2.5
mysql-connector-python==8.0.33
protobuf==3.20.3
pycryptodome==3.18.0
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tzlocal==5.0.1
urllib3==1.26.16
Werkzeug==2.3.4
zipp==3.15.0
